# learn-and-reward-lar/docs/contracts/openapi-mission-b.json
{"openapi":"3.1.0","info":{"title":"Learn-and-Reward-LAR API","description":"Modular monolith API. Mission B scope: learning + ai_chat modules (Developer B). Auth is a placeholder for Developer C's core module.","version":"0.1.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["auth"],"summary":"Read Current User","operationId":"read_current_user_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/courses":{"get":{"tags":["learning"],"summary":"List Courses","operationId":"list_courses_api_v1_courses_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response List Courses Api V1 Courses Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/courses/{course_id}":{"get":{"tags":["learning"],"summary":"Get Course","operationId":"get_course_api_v1_courses__course_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/courses/{course_id}/enroll":{"post":{"tags":["learning"],"summary":"Enroll Course","operationId":"enroll_course_api_v1_courses__course_id__enroll_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"course_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrollResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/lessons/{lesson_id}":{"get":{"tags":["learning"],"summary":"Get Lesson","operationId":"get_lesson_api_v1_lessons__lesson_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"lesson_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Lesson Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/practice/sessions":{"post":{"tags":["learning"],"summary":"Create Practice Session","operationId":"create_practice_session_api_v1_practice_sessions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeSessionCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeSessionDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/practice/sessions/{session_id}":{"get":{"tags":["learning"],"summary":"Get Practice Session","operationId":"get_practice_session_api_v1_practice_sessions__session_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PracticeSessionDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/practice/sessions/{session_id}/messages":{"post":{"tags":["learning"],"summary":"Submit Session Message","description":"Unified session message endpoint.\n\n* ``lesson_exercises`` sessions take an :class:`ExerciseSubmission`.\n* ``free_conversation`` sessions take a :class:`FreeChatMessage` and get an\n  AI reply (RAG context + DeepSeek) through the ai_chat module.","operationId":"submit_session_message_api_v1_practice_sessions__session_id__messages_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/ExerciseSubmission"},{"$ref":"#/components/schemas/FreeChatMessage"}],"title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/SubmitResponse"},{"$ref":"#/components/schemas/AiChatResponse"}],"title":"Response Submit Session Message Api V1 Practice Sessions  Session Id  Messages Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ai/chat":{"post":{"tags":["learning"],"summary":"Ai Chat","operationId":"ai_chat_api_v1_ai_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AiChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AiChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/certificates/claim":{"post":{"tags":["learning"],"summary":"Claim Certificate","operationId":"claim_certificate_api_v1_certificates_claim_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CertificateClaimRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CertificateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/certificates":{"get":{"tags":["learning"],"summary":"List My Certificates","operationId":"list_my_certificates_api_v1_certificates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CertificateResponse"},"type":"array","title":"Response List My Certificates Api V1 Certificates Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/certificates/{code}/verify":{"get":{"tags":["learning"],"summary":"Verify Certificate","operationId":"verify_certificate_api_v1_certificates__code__verify_get","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CertificateVerifyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"AiChatRequest":{"properties":{"message":{"type":"string","maxLength":4000,"minLength":1,"title":"Message"},"session_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Session Id","description":"Existing practice session to continue; a new one is created when omitted."}},"type":"object","required":["message"],"title":"AiChatRequest"},"AiChatResponse":{"properties":{"session_id":{"type":"string","format":"uuid","title":"Session Id"},"reply":{"type":"string","title":"Reply"},"citations":{"items":{"$ref":"#/components/schemas/Citation"},"type":"array","title":"Citations"},"disclaimer":{"type":"string","title":"Disclaimer"}},"type":"object","required":["session_id","reply","citations","disclaimer"],"title":"AiChatResponse"},"CertificateClaimRequest":{"properties":{"course_id":{"type":"string","format":"uuid","title":"Course Id"}},"type":"object","required":["course_id"],"title":"CertificateClaimRequest"},"CertificateResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"code":{"type":"string","title":"Code"},"course_id":{"type":"string","format":"uuid","title":"Course Id"},"title":{"type":"string","title":"Title"},"issued_at":{"type":"string","format":"date-time","title":"Issued At"},"revoked":{"type":"boolean","title":"Revoked"}},"type":"object","required":["id","code","course_id","title","issued_at","revoked"],"title":"CertificateResponse"},"CertificateVerifyResponse":{"properties":{"valid":{"type":"boolean","title":"Valid"},"code":{"type":"string","title":"Code"},"title":{"type":"string","title":"Title"},"issued_at":{"type":"string","format":"date-time","title":"Issued At"},"revoked":{"type":"boolean","title":"Revoked"}},"type":"object","required":["valid","code","title","issued_at","revoked"],"title":"CertificateVerifyResponse"},"ChatMessageResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"session_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Session Id"},"role":{"$ref":"#/components/schemas/MessageRole"},"content":{"type":"string","title":"Content"},"citations":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Citations"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","session_id","role","content","created_at"],"title":"ChatMessageResponse"},"Citation":{"properties":{"index":{"type":"integer","title":"Index"},"chunk_id":{"type":"string","title":"Chunk Id"},"resource_id":{"type":"string","title":"Resource Id"},"snippet":{"type":"string","title":"Snippet"}},"type":"object","required":["index","chunk_id","resource_id","snippet"],"title":"Citation"},"CourseDetail":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"slug":{"type":"string","title":"Slug"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"language":{"type":"string","title":"Language"},"level":{"$ref":"#/components/schemas/CourseLevel"},"price_coins":{"type":"integer","title":"Price Coins"},"is_published":{"type":"boolean","title":"Is Published"},"lesson_count":{"type":"integer","title":"Lesson Count"},"lessons":{"items":{"$ref":"#/components/schemas/LessonSummary"},"type":"array","title":"Lessons","default":[]},"enrollment":{"anyOf":[{"$ref":"#/components/schemas/EnrollmentSummary"},{"type":"null"}]}},"type":"object","required":["id","slug","title","description","language","level","price_coins","is_published","lesson_count"],"title":"CourseDetail"},"CourseLevel":{"type":"string","enum":["beginner","intermediate","advanced"],"title":"CourseLevel"},"EnrollResponse":{"properties":{"enrollment":{"$ref":"#/components/schemas/EnrollmentSummary"},"coins_spent":{"type":"integer","title":"Coins Spent"}},"type":"object","required":["enrollment","coins_spent"],"title":"EnrollResponse"},"EnrollmentStatus":{"type":"string","enum":["active","completed"],"title":"EnrollmentStatus"},"EnrollmentSummary":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"course_id":{"type":"string","format":"uuid","title":"Course Id"},"status":{"$ref":"#/components/schemas/EnrollmentStatus"},"lessons_completed":{"type":"integer","title":"Lessons Completed"},"total_lessons":{"type":"integer","title":"Total Lessons"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","course_id","status","lessons_completed","total_lessons","created_at"],"title":"EnrollmentSummary"},"ExerciseFeedback":{"properties":{"correct":{"type":"boolean","title":"Correct"},"expected":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Expected"},"explanation":{"type":"string","title":"Explanation","default":""},"ai_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ai Hint"}},"type":"object","required":["correct"],"title":"ExerciseFeedback"},"ExercisePublic":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"order_index":{"type":"integer","title":"Order Index"},"type":{"$ref":"#/components/schemas/ExerciseType"},"prompt":{"type":"string","title":"Prompt"},"options":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Options"},"points":{"type":"integer","title":"Points"}},"type":"object","required":["id","order_index","type","prompt","points"],"title":"ExercisePublic"},"ExerciseSubmission":{"properties":{"exercise_id":{"type":"string","format":"uuid","title":"Exercise Id"},"answer":{"anyOf":[{"type":"string"},{"type":"integer"}],"title":"Answer","description":"Option index for multiple_choice, text otherwise."}},"type":"object","required":["exercise_id","answer"],"title":"ExerciseSubmission"},"ExerciseType":{"type":"string","enum":["multiple_choice","fill_in_blank","translation"],"title":"ExerciseType"},"FreeChatMessage":{"properties":{"content":{"type":"string","maxLength":4000,"minLength":1,"title":"Content"}},"type":"object","required":["content"],"title":"FreeChatMessage"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LessonDetail":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"course_id":{"type":"string","format":"uuid","title":"Course Id"},"course_title":{"type":"string","title":"Course Title"},"order_index":{"type":"integer","title":"Order Index"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"vocabulary":{"items":{"$ref":"#/components/schemas/VocabularyItem"},"type":"array","title":"Vocabulary","default":[]},"exercises":{"items":{"$ref":"#/components/schemas/ExercisePublic"},"type":"array","title":"Exercises","default":[]}},"type":"object","required":["id","course_id","course_title","order_index","title","content"],"title":"LessonDetail"},"LessonSummary":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"order_index":{"type":"integer","title":"Order Index"},"title":{"type":"string","title":"Title"},"completed":{"type":"boolean","title":"Completed","default":false}},"type":"object","required":["id","order_index","title"],"title":"LessonSummary"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MessageRole":{"type":"string","enum":["user","assistant","system"],"title":"MessageRole"},"PracticeKind":{"type":"string","enum":["lesson_exercises","free_conversation"],"title":"PracticeKind"},"PracticeSessionCreate":{"properties":{"kind":{"$ref":"#/components/schemas/PracticeKind","default":"lesson_exercises"},"lesson_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Lesson Id"}},"type":"object","title":"PracticeSessionCreate"},"PracticeSessionDetail":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"kind":{"$ref":"#/components/schemas/PracticeKind"},"status":{"$ref":"#/components/schemas/SessionStatus"},"lesson_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Lesson Id"},"total_exercises":{"type":"integer","title":"Total Exercises"},"correct_exercises":{"type":"integer","title":"Correct Exercises"},"coins_awarded":{"type":"integer","title":"Coins Awarded"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"exercises":{"items":{"$ref":"#/components/schemas/ExercisePublic"},"type":"array","title":"Exercises","default":[]},"messages":{"items":{"$ref":"#/components/schemas/ChatMessageResponse"},"type":"array","title":"Messages","default":[]}},"type":"object","required":["id","kind","status","lesson_id","total_exercises","correct_exercises","coins_awarded","created_at"],"title":"PracticeSessionDetail"},"PracticeSessionSummary":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"kind":{"$ref":"#/components/schemas/PracticeKind"},"status":{"$ref":"#/components/schemas/SessionStatus"},"lesson_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Lesson Id"},"total_exercises":{"type":"integer","title":"Total Exercises"},"correct_exercises":{"type":"integer","title":"Correct Exercises"},"coins_awarded":{"type":"integer","title":"Coins Awarded"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","kind","status","lesson_id","total_exercises","correct_exercises","coins_awarded","created_at"],"title":"PracticeSessionSummary"},"RegisterRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"},"display_name":{"type":"string","maxLength":120,"minLength":1,"title":"Display Name"}},"type":"object","required":["email","password","display_name"],"title":"RegisterRequest"},"SessionStatus":{"type":"string","enum":["in_progress","completed"],"title":"SessionStatus"},"SubmitResponse":{"properties":{"message_id":{"type":"string","format":"uuid","title":"Message Id"},"feedback":{"anyOf":[{"$ref":"#/components/schemas/ExerciseFeedback"},{"type":"null"}]},"reply":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reply"},"citations":{"items":{},"type":"array","title":"Citations","default":[]},"session":{"$ref":"#/components/schemas/PracticeSessionSummary"},"coins_awarded":{"type":"integer","title":"Coins Awarded","default":0}},"type":"object","required":["message_id","feedback","session"],"title":"SubmitResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"TokenResponse"},"UserResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","title":"Email"},"display_name":{"type":"string","title":"Display Name"},"role":{"type":"string","title":"Role"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","display_name","role","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VocabularyItem":{"properties":{"term":{"type":"string","title":"Term"},"translation":{"type":"string","title":"Translation"},"pronunciation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pronunciation"}},"type":"object","required":["term","translation"],"title":"VocabularyItem"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}